// app/enums/Screen.ts
export enum Screen {
    DASHBOARD = 'DASHBOARD',
    ENTRY_LIST = 'ENTRY_LIST',
    ENTRY_FORM = 'ENTRY_FORM',
    ENTRY_DETAILS = 'ENTRY_DETAILS',
    ENTRIES_FILTER = 'ENTRIES_FILTER',
    TROPHY_LIST = 'TROPHY_LIST',
    TROPHY_FORM = 'TROPHY_FORM',
    TROPHY_DETAILS = 'TROPHY_DETAILS',
    TROPHIES_FILTER = 'TROPHIES_FILTER',
    TIMELINE = 'TIMELINE',
    TIMELINE_DETAILS = 'TIMELINE_DETAILS',
    STATISTICS = 'STATISTICS',
    SETTINGS = 'SETTINGS'
}

export enum TabScreen {
    DASHBOARD = 'Dashboard',
    TIMELINE = 'Timeline',
    STATISTICS = 'Statistics',
    SETTINGS = 'Settings'
}